Deploy ArkSwap V1 to Ark Constellation devnet (chain id 9000)
Ran the full llm.txt section 26 deployment order against the live devnet and
recorded the results.

Deployed and verified on Blockscout
- ArkSwapFactory  0x594f74aDCa63Af06d10Ee06fEe8F237DBb560e06 (feeTo = address(0))
- ArkSwapRouter02 0xF2a7fD4072d70D2e7eDF5a5eB653965CEe0871b6
- mUSDC           0xC8d15f9A42Ee3107ab1257E38199e3Bf899dCfB3
- mUSDT           0x073D643F712C136D2e7F71d683d9E26b6EFC845b

Pointed at the canonical WKASH 0x6792d2fD02d8A55c543F627d0e90526f9278C6d6,
after checking on-chain that symbol() is WKASH and decimals() is 18.

Live checks
- ArkSwapLibrary.pairFor derivation matches the deployed factory for all 3 pairs.
- KASH -> mUSDC and mUSDC -> KASH swaps; k grew by the fee on the round trip.
- Multi-hop mUSDT -> WKASH -> mUSDC delivered exactly the quoted amount.
- 25% liquidity removal returned principal plus accrued fees.
- Frontend reads live reserves; its quote matches the router exactly.

Deployment targets now run with --slow --gas-estimate-multiplier 300. Ark's
in-simulation gas estimates assume warm storage and under-estimated a swap by
about 25k gas (115,976 estimated vs ~140,092 needed). The shortfall left the
pair's SSTORE in _update() with under 2300 gas and tripped the EIP-2200
reentrancy sentry, surfacing as ReentrancySentryOOG rather than a plain
out-of-gas. Resending with adequate gas succeeded; this is not a V2/Ark
incompatibility.

Pair contracts are live but not yet verified: this Blockscout instance does not
index CREATE2 creations from internal transactions, so it does not recognise the
pair addresses as contracts. Retry with `make verify-pairs`.

Adds docs/PRODUCTION-READINESS.md recording what is done and the gates that
remain open. This deployment is devnet only and is not production-ready: it has
not been audited, feeToSetter is a single EOA, and repository licensing is still
unresolved.

+++ script/LocalPreview.s.sol
@@ -1,34 +1,34 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity ^0.8.24;
 
-import {ArkScript} from './ArkScript.sol';
-import {console2} from 'forge-std/Script.sol';
+import {ArkScript} from "./ArkScript.sol";
+import {console2} from "forge-std/Script.sol";
 
 interface ITokenOps {
-    function approve(address, uint) external returns (bool);
-    function mint(address, uint) external;
+    function approve(address, uint256) external returns (bool);
+    function mint(address, uint256) external;
 }
 
 interface IRouterSeed {
     function addLiquidityETH(
         address token,
-        uint amountTokenDesired,
-        uint amountTokenMin,
-        uint amountETHMin,
+        uint256 amountTokenDesired,
+        uint256 amountTokenMin,
+        uint256 amountETHMin,
         address to,
-        uint deadline
-    ) external payable returns (uint, uint, uint);
+        uint256 deadline
+    ) external payable returns (uint256, uint256, uint256);
 
     function addLiquidity(
         address tokenA,
         address tokenB,
-        uint amountADesired,
-        uint amountBDesired,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint, uint, uint);
+        uint256 deadline
+    ) external returns (uint256, uint256, uint256);
 }
 
 /// @notice Stands up a COMPLETE ArkSwap stack on a LOCAL dev chain, for frontend
@@ -46,56 +46,50 @@ interface IRouterSeed {
 ///          --rpc-url http://127.0.0.1:8545 --broadcast \
 ///          --private-key 0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
 contract LocalPreview is ArkScript {
-    uint internal constant ANVIL_CHAIN_ID = 31337;
+    uint256 internal constant ANVIL_CHAIN_ID = 31337;
 
     function run() external {
         require(
             block.chainid == ANVIL_CHAIN_ID,
-            'LocalPreview: local dev chain only -- this deploys a WKASH test double and must never touch Ark'
+            "LocalPreview: local dev chain only -- this deploys a WKASH test double and must never touch Ark"
         );
 
-        uint pk = vm.envOr('LOCAL_PRIVATE_KEY', uint(0));
-        require(pk != 0, 'LocalPreview: set LOCAL_PRIVATE_KEY');
+        uint256 pk = vm.envOr("LOCAL_PRIVATE_KEY", uint256(0));
+        require(pk != 0, "LocalPreview: set LOCAL_PRIVATE_KEY");
         address me = vm.addr(pk);
 
         vm.startBroadcast(pk);
 
-        address wkash = deployCode('WKASH9.sol:WKASH9');
-        address factory = deployCode('ArkSwapFactory.sol:ArkSwapFactory', abi.encode(me));
-        address router = deployCode('ArkSwapRouter02.sol:ArkSwapRouter02', abi.encode(factory, wkash));
-        address musdc = deployCode('MockUSDC.sol:MockUSDC', abi.encode(uint(0)));
-        address musdt = deployCode('MockUSDT.sol:MockUSDT', abi.encode(uint(0)));
+        address wkash = deployCode("WKASH9.sol:WKASH9");
+        address factory = deployCode("ArkSwapFactory.sol:ArkSwapFactory", abi.encode(me));
+        address router = deployCode("ArkSwapRouter02.sol:ArkSwapRouter02", abi.encode(factory, wkash));
+        address musdc = deployCode("MockUSDC.sol:MockUSDC", abi.encode(uint256(0)));
+        address musdt = deployCode("MockUSDT.sol:MockUSDT", abi.encode(uint256(0)));
 
         // Devnet reference pricing: 1 KASH = 1 mUSDC (llm.txt s16).
         ITokenOps(musdc).mint(me, 500_000e6);
         ITokenOps(musdt).mint(me, 500_000e6);
-        ITokenOps(musdc).approve(router, type(uint).max);
-        ITokenOps(musdt).approve(router, type(uint).max);
+        ITokenOps(musdc).approve(router, type(uint256).max);
+        ITokenOps(musdt).approve(router, type(uint256).max);
 
-        IRouterSeed(router).addLiquidityETH{value: 2_000 ether}(
-            musdc, 2_000e6, 0, 0, me, block.timestamp + 1 hours
-        );
-        IRouterSeed(router).addLiquidityETH{value: 1_000 ether}(
-            musdt, 1_000e6, 0, 0, me, block.timestamp + 1 hours
-        );
-        IRouterSeed(router).addLiquidity(
-            musdc, musdt, 100_000e6, 100_000e6, 0, 0, me, block.timestamp + 1 hours
-        );
+        IRouterSeed(router).addLiquidityETH{value: 2_000 ether}(musdc, 2_000e6, 0, 0, me, block.timestamp + 1 hours);
+        IRouterSeed(router).addLiquidityETH{value: 1_000 ether}(musdt, 1_000e6, 0, 0, me, block.timestamp + 1 hours);
+        IRouterSeed(router).addLiquidity(musdc, musdt, 100_000e6, 100_000e6, 0, 0, me, block.timestamp + 1 hours);
 
         vm.stopBroadcast();
 
-        string memory out = 'preview';
-        vm.serializeAddress(out, 'wkash', wkash);
-        vm.serializeAddress(out, 'factory', factory);
-        vm.serializeAddress(out, 'router', router);
-        vm.serializeAddress(out, 'musdc', musdc);
-        string memory json = vm.serializeAddress(out, 'musdt', musdt);
-        vm.writeJson(json, './deployments/local-preview.json');
+        string memory out = "preview";
+        vm.serializeAddress(out, "wkash", wkash);
+        vm.serializeAddress(out, "factory", factory);
+        vm.serializeAddress(out, "router", router);
+        vm.serializeAddress(out, "musdc", musdc);
+        string memory json = vm.serializeAddress(out, "musdt", musdt);
+        vm.writeJson(json, "./deployments/local-preview.json");
 
-        console2.log('WKASH9  ', wkash);
-        console2.log('FACTORY ', factory);
-        console2.log('ROUTER  ', router);
-        console2.log('mUSDC   ', musdc);
-        console2.log('mUSDT   ', musdt);
+        console2.log("WKASH9  ", wkash);
+        console2.log("FACTORY ", factory);
+        console2.log("ROUTER  ", router);
+        console2.log("mUSDC   ", musdc);
+        console2.log("mUSDT   ", musdt);
     }
 }
